增加增量访问及类及方法注解访问-1

+++ src/main/groovy/com/qiyei/android/asmplugin/AsmTransform.groovy
@@ -9,14 +9,20 @@ import com.android.build.api.transform.TransformException
 import com.android.build.api.transform.TransformInput
 import com.android.build.api.transform.TransformInvocation
 import com.android.build.api.transform.TransformOutputProvider
+import com.android.build.api.transform.Status
 import com.android.build.gradle.internal.pipeline.TransformManager
+import com.android.ide.common.internal.WaitableExecutor
 import com.android.utils.FileUtils
 import org.apache.commons.codec.digest.DigestUtils
 import org.apache.commons.compress.utils.IOUtils
 import org.objectweb.asm.ClassReader
 import org.objectweb.asm.ClassVisitor
 import org.objectweb.asm.ClassWriter
+import org.objectweb.asm.tree.AnnotationNode
+import org.objectweb.asm.tree.ClassNode
+import org.objectweb.asm.tree.MethodNode
 
+import java.util.concurrent.Callable
 import java.util.jar.JarEntry
 import java.util.jar.JarFile
 import java.util.jar.JarOutputStream
@@ -24,6 +30,7 @@ import java.util.zip.ZipEntry
 
 
 public class AsmTransform extends Transform {
+    private WaitableExecutor mWaitableExecutor = WaitableExecutor.useGlobalSharedThreadPool()
 
     /**
      * 设置我们自定义的Transform对应的Task名称
@@ -56,7 +63,7 @@ public class AsmTransform extends Transform {
 
     @Override
     boolean isIncremental() {
-        return false
+        return true
     }
 
     @Override
@@ -66,7 +73,13 @@ public class AsmTransform extends Transform {
         //拿到所有的class文件
         Collection<TransformInput> inputs = transformInvocation.inputs
         TransformOutputProvider outputProvider = transformInvocation.outputProvider
-        if (outputProvider != null) {
+
+        //当前是否是增量编译,由isIncremental方法决定的
+        // 当上面的isIncremental()写的返回true,这里得到的值不一定是true,还得看当时环境.比如clean之后第一次运行肯定就不是增量编译嘛.
+        boolean isIncremental = transformInvocation.isIncremental()
+        println("transform isIncremental = $isIncremental")
+        if (!isIncremental && outputProvider != null) {
+            //不是增量编译则删除之前的所有文件
             outputProvider.deleteAll()
         }
         //遍历inputs Transform的inputs有两种类型，一种是目录，一种是jar包，要分开遍历
@@ -75,15 +88,32 @@ public class AsmTransform extends Transform {
 
             // 比如我们手写的类以及R.class、BuildConfig.class以及R$XXX.class等
             input.directoryInputs.each { DirectoryInput directoryInput ->
-                //文件夹中的class文件
-                handDirectoryInput(directoryInput, outputProvider)
+                //多线程
+                mWaitableExecutor.execute(new Callable<Object>() {
+                    @Override
+                    Object call() throws Exception {
+                        //文件夹中的class文件
+                        processSourceInput(directoryInput, outputProvider, isIncremental)
+                        return null
+                    }
+                })
             }
 
             //遍历jar包中的class文件 jarInputs代表以jar包方式参与项目编译的所有本地jar包或远程jar包
             input.jarInputs.each { JarInput jarInput ->
-                //处理jar包中的class文件
-                //handJarInput(jarInput, outputProvider)
+                //多线程处理jar
+                mWaitableExecutor.execute(new Callable<Object>() {
+                    @Override
+                    Object call() throws Exception {
+                        //处理jar包中的class文件
+                        //processJarInput(jarInput, outputProvider,isIncremental)
+                        return null
+                    }
+                })
+
             }
+            //等待所有任务结束
+            mWaitableExecutor.waitForTasksWithQuickFail(true)
         }
     }
 
@@ -92,39 +122,139 @@ public class AsmTransform extends Transform {
      * @param input
      * @param outputProvider
      */
-    private static void handDirectoryInput(DirectoryInput input, TransformOutputProvider outputProvider) {
+    private void processSourceInput(DirectoryInput input, TransformOutputProvider outputProvider, boolean isIncremental) {
+        File dest = outputProvider.getContentLocation(input.name, input.contentTypes, input.scopes, Format.DIRECTORY)
+        FileUtils.forceMkdir(dest)
+
+        if (isIncremental) {
+            String srcDirPath = input.getFile().getAbsolutePath()
+            String destDirPath = dest.getAbsolutePath()
+            Map<File, Status> fileStatusMap = input.getChangedFiles()
+            for (Map.Entry<File, Status> changedFile : fileStatusMap.entrySet()) {
+                Status status = changedFile.getValue()
+                File inputFile = changedFile.getKey()
+                String destFilePath = inputFile.getAbsolutePath().replace(srcDirPath, destDirPath)
+                File destFile = new File(destFilePath)
+                switch (status) {
+                    case Status.NOTCHANGED:
+                        break
+                    case Status.ADDED:
+                    case Status.CHANGED:
+                        FileUtils.touch(destFile)
+                        interceptDirClass(inputFile)
+                        transformSingleFile(inputFile, destFile)
+                        break
+                    case Status.REMOVED:
+                        if (destFile.exists()) {
+                            FileUtils.forceDelete(destFile)
+                        }
+                        break
+                }
+            }
+        } else {
+            interceptDirClass(input.file)
+            transformDirectory(input.file, dest)
+        }
+    }
+
+
+    private void transformSingleFile(File inputFile, File destFile) {
+        FileUtils.copyFile(inputFile, destFile)
+    }
+
+    private void transformDirectory(File directoryInputFile, File dest) {
+        FileUtils.copyDirectory(directoryInputFile, dest)
+    }
+
+    private void interceptDirClass(JarInput input){
         //是否是文件夹
         if (input.file.isDirectory()) {
             //列出目录所有文件（包含子文件夹，子文件夹内文件）
             input.file.eachFileRecurse { File file ->
-                String name = file.name
-                //需要插桩class 根据自己的需求来------------- 这里判断是否是我们自己写的Application
-                println("handDirectoryInput=" + name)
-                if ("VisionApplication.class".equals(name)) {
-                    ClassReader classReader = new ClassReader(file.bytes)
-                    //传入COMPUTE_MAXS  ASM会自动计算本地变量表和操作数栈
-                    ClassWriter classWriter = new ClassWriter(classReader, ClassWriter.COMPUTE_MAXS)
-                    //创建类访问器   并交给它去处理
-                    ClassVisitor classVisitor = new LogVisitor(classWriter)
-                    classReader.accept(classVisitor, ClassReader.EXPAND_FRAMES)
-                    byte[] code = classWriter.toByteArray()
-                    FileOutputStream fos = new FileOutputStream(file.parentFile.absolutePath + File.separator + name)
-                    fos.write(code)
-                    fos.close()
+                visitorClass(file)
+            }
+        } else {
+            visitorClass(input.file)
+        }
+    }
+
+    private void visitorClass(File file){
+        String name = file.name
+        //需要插桩class 根据自己的需求来------------- 这里判断是否是我们自己写的Application
+        println("visitorClass=" + name)
+        ClassReader classReader = new ClassReader(file.bytes)
+        //创建ClassNode,读取的信息会封装到这个类里面
+        ClassNode cn = new ClassNode()
+        //开始读取
+        classReader.accept(cn, 0)
+        //获取声明的所有注解
+        List<AnnotationNode> annotations = cn.visibleAnnotations
+        if(annotations!=null) {//便利注解
+            for(AnnotationNode an: annotations) {
+                //获取注解的描述信息
+                String anno = an.desc.replaceAll("/", ".");
+                String annoName = anno.substring(1, anno.length()-1);
+                if("com.mh.base.quartz.annotation.BaseQuartz".equals(annoName)) {
+                    String className = cn.name.replaceAll("/", ".");
+                    //获取注解的属性名对应的值，（values是一个集合，它将注解的属性和属性值都放在了values中，通常奇数为值偶数为属性名）
+                    String valu = an.values.get(1).toString();
+                    System.out.println(className);
+                    System.out.println(valu);
+                    //根据匹配的注解，将其封装给具体的业务使用
+                    //taskClazz.put(valu, Class.forName(className));
                 }
             }
         }
-        //处理完输入文件后把输出传给下一个文件
-        def dest = outputProvider.getContentLocation(input.name, input.contentTypes, input.scopes, Format.DIRECTORY)
-        FileUtils.copyDirectory(input.file, dest)
+
+        //传入COMPUTE_MAXS  ASM会自动计算本地变量表和操作数栈
+        ClassWriter classWriter = new ClassWriter(classReader, ClassWriter.COMPUTE_MAXS)
+        //创建类访问器   并交给它去处理
+        ClassVisitor classVisitor = new CustomInterceptVisitor(classWriter)
+        classReader.accept(classVisitor, ClassReader.EXPAND_FRAMES)
+        byte[] code = classWriter.toByteArray()
+        FileOutputStream fos = new FileOutputStream(file.parentFile.absolutePath + File.separator + name)
+        fos.write(code)
+        fos.close()
     }
+
+    private void tetst(){
+
+    }
+
 
     /**
      * 遍历jarInputs 得到对应的class 交给ASM处理
      * @param jarInput
      * @param outputProvider
      */
-    private static void handJarInput(JarInput jarInput, TransformOutputProvider outputProvider) {
+    private void processJarInput(JarInput jarInput, TransformOutputProvider outputProvider, boolean isIncremental) {
+        def status = jarInput.status
+        File dest = outputProvider.getContentLocation(jarInput.file.absolutePath, jarInput.contentTypes, jarInput.scopes, Format.JAR)
+        if (isIncremental) {
+            switch (status) {
+                case Status.NOTCHANGED:
+                    break
+                case Status.ADDED:
+                case Status.CHANGED:
+                    transformJar(interceptJar(jarInput), dest)
+                    break
+                case Status.REMOVED:
+                    if (dest.exists()) {
+                        FileUtils.forceDelete(dest)
+                    }
+                    break
+            }
+        } else {
+            transformJar(interceptJar(jarInput), dest)
+        }
+    }
+
+    private void transformJar(File jarInputFile, File dest) {
+        FileUtils.copyFile(jarInputFile, dest)
+    }
+
+    private File interceptJar(JarInput jarInput){
+        File tmpFile = null
         if (jarInput.file.getAbsolutePath().endsWith(".jar")) {
             //重名名输出文件,因为可能同名,会覆盖
             def jarName = jarInput.name
@@ -134,7 +264,7 @@ public class AsmTransform extends Transform {
             }
             JarFile jarFile = new JarFile(jarInput.file)
             Enumeration enumeration = jarFile.entries()
-            File tmpFile = new File(jarInput.file.getParent() + File.separator + "classes_temp.jar")
+            tmpFile = new File(jarInput.file.getParent() + File.separator + "classes_temp.jar")
             //避免上次的缓存被重复插入
             if (tmpFile.exists()) {
                 tmpFile.delete()
@@ -155,7 +285,7 @@ public class AsmTransform extends Transform {
                     ClassReader classReader = new ClassReader(IOUtils.toByteArray(inputStream))
                     ClassWriter classWriter = new ClassWriter(classReader, ClassWriter.COMPUTE_MAXS)
                     //创建类访问器   并交给它去处理
-                    ClassVisitor cv = new LogVisitor(classWriter)
+                    ClassVisitor cv = new CustomInterceptVisitor(classWriter)
                     classReader.accept(cv, ClassReader.EXPAND_FRAMES)
                     byte[] code = classWriter.toByteArray()
                     jarOutputStream.write(code)
@@ -168,11 +298,9 @@ public class AsmTransform extends Transform {
             //结束
             jarOutputStream.close()
             jarFile.close()
-            //获取output目录
-            def dest = outputProvider.getContentLocation(jarName + md5Name,
-                    jarInput.contentTypes, jarInput.scopes, Format.JAR)
-            FileUtils.copyFile(tmpFile, dest)
-            tmpFile.delete()
+        } else {
+            tmpFile = jarInput.file
         }
+        return tmpFile
     }
 }
